Refactor rule conditions
