+++ project/rule.go
@@ -14,7 +14,6 @@
 package project
 
 import (
-	"context"
 	"fmt"
 	"path"
 	"strings"
@@ -386,60 +385,6 @@ func (r *Rule) DependencyOutputs() (outputs Resources) {
 		}
 	}
 	return
-}
-
-// CheckConditions returns true if the Rule should execute based on its conditions
-func (r *Rule) CheckConditions(ctx context.Context, executor Executor, debug bool) (bool, error) {
-	if !r.when.IsEmpty() {
-		// A when condition is defined
-		whenCondition, err := r.checkCondition(ctx, executor, debug, r.when)
-		if err != nil {
-			return false, err
-		}
-		if !whenCondition {
-			// The "when" condition evaluted to false: condition not met
-			return false, nil
-		}
-	}
-	if !r.unless.IsEmpty() {
-		// An unless condition is defined
-		unlessCondition, err := r.checkCondition(ctx, executor, debug, r.unless)
-		if err != nil {
-			return false, err
-		}
-		if unlessCondition {
-			// The "unless" condition evaluted to true: condition not met
-			return false, nil
-		}
-	}
-	// All conditions met
-	return true, nil
-}
-
-func (r *Rule) checkCondition(ctx context.Context, executor Executor, debug bool, c Condition) (bool, error) {
-	if c.ResourceExists != "" {
-		// The "resource exists" condition evaluates to true if one or more resources
-		// match the provided filename or glob pattern
-		resources, err := matchResources(r.Component(), r.inProvider, []string{c.ResourceExists})
-		if err != nil {
-			return false, err
-		}
-		return len(resources) > 0, nil
-	}
-	if c.ScriptSucceeds != "" {
-		// The "script succeeds" condition evaluates to true if the specified shell
-		// command exits without error when run in bash
-		err := executor.Execute(ctx, ExecOpts{
-			Command:          c.ScriptSucceeds,
-			WorkingDirectory: r.Component().Directory(),
-			Env:              flattenEnvironment(r.BaseEnvironment()),
-			Image:            r.Image(),
-			Name:             fmt.Sprintf("%s.condition", r.NodeID()),
-			Debug:            debug,
-		})
-		return err == nil, nil
-	}
-	return true, nil
 }
 
 // Commands that define Rule execution
